`monitor`: `parallel` option
Adds `parallel?: number` to `monitor` options.
This option configures how much tasks could run in parallel.
